Deliverydate in Header or Lines
The deliverydate is requrired in header.
Alternatively a deliveryperiod
Currently the invoicedate is used for deliverydate.
This is okay, if the delivery and invoice has the same date.
In the most cases, the invoice have a later date.
So is possible the invoice will be rejected.
Also by consolidated invoices.
For consolidatet invoices the alternativ field deliveryperiod can
be used, in header and in the lines for products(services) and
charges.
resolve #38